XIVY-18535 handle openapi file-pick request in integration
